implement DrawMask
Drawmask was not implemented before.
Edwood uses it with an image allocated with the repl bit
and larger than a single pixel, which received an image.Uniform.
But uniform images are not drawable.
We now send an image.RGBA and ignore the repl bit, if the size
is larger than one pixel.

For rjkroege/edwood#116