change package name from draw to duitdraw
The package was called draw initially, to be a
drop-in replacement for 9fans.net/go/draw.
This causes confusion and strange error messages.
It makes it also more complicated to import image/draw together with
duitdraw.
Both ktye/duit and edwood based on duitdraw import duitdraw as draw
anyway, so this should not break things.