Use primary selection in X11 for snarf read/write
We try to match the behavior of devdraw. Devdraw writes to both the
clipboard and the primary selection in X11.  For reads, it first tries
to read the primary selection, and falls back to the clipboard if
it's empty.

We use a slightly modified version of the X11 clipboard implementation
in nucular.

I tried to use atotto/clipboard for access to primary selection, but
it seems to broken, has weird API where you need to set package level
variable `Primary` to `true`, and depends on the `xclip` or `xsel`
command being installed in the system.

Using nucular/clipboard directly is an option, but it has poor error
handling and read buffer seems to be fixed at 5MB.